FIX: fixed warnings

+++ Core/DataStructures/mitkSeedsImage.cpp
@@ -163,10 +163,10 @@ void mitk::SeedsImage::AddSeedPoint(SeedsImageType* itkImage)
     for(unsigned int i=0; i<dimension; i++)
       radius[i] = int(m_Radius/spacing[i]);
     //FillVector2D(radius, ((ScalarType)m_Radius)/spacing[0], ((ScalarType)m_Radius)/spacing[1]);
-      for(int y = contIndex[1] - radius[1]; y <= contIndex[1] + radius[1]; ++y){
-        for(int x = contIndex[0] - radius[0]; x <= contIndex[0] + radius[0]; ++x){
-          delta_x = abs(x - contIndex[0])*spacing[0];
-          delta_y = abs(y - contIndex[1])*spacing[1];
+      for(int y = (int)contIndex[1] - radius[1]; y <= (int)contIndex[1] + radius[1]; ++y){
+        for(int x = (int)contIndex[0] - radius[0]; x <= (int)contIndex[0] + radius[0]; ++x){
+          delta_x = abs(x - (int)contIndex[0])*spacing[0];
+          delta_y = abs(y - (int)contIndex[1])*spacing[1];
           sphere_distance = (delta_x * delta_x) + (delta_y * delta_y);
           if (sphere_distance <= (m_Radius*m_Radius)){
             // check -> is the point inside the image?
@@ -187,12 +187,12 @@ void mitk::SeedsImage::AddSeedPoint(SeedsImageType* itkImage)
     int radius[dimension];
     for(unsigned int i=0; i<dimension; i++)
       radius[i] = int(m_Radius/spacing[i]);
-    for(int z = contIndex[2] - radius[2]; z <= contIndex[2] + radius[2]; ++z){
-      for(int y = contIndex[1] - radius[1]; y <= contIndex[1] + radius[1]; ++y){
-        for(int x = contIndex[0] - radius[0]; x <= contIndex[0] + radius[0]; ++x){
-          delta_x = abs(x - contIndex[0])*spacing[0];
-          delta_y = abs(y - contIndex[1])*spacing[1];
-          delta_z = abs(z - contIndex[2])*spacing[2];
+    for(int z = (int)contIndex[2] - radius[2]; z <= (int)contIndex[2] + radius[2]; ++z){
+      for(int y = (int)contIndex[1] - radius[1]; y <= (int)contIndex[1] + radius[1]; ++y){
+        for(int x = (int)contIndex[0] - radius[0]; x <= (int)contIndex[0] + radius[0]; ++x){
+          delta_x = abs(x - (int)contIndex[0])*spacing[0];
+          delta_y = abs(y - (int)contIndex[1])*spacing[1];
+          delta_z = abs(z - (int)contIndex[2])*spacing[2];
           sphere_distance = (delta_x * delta_x) + (delta_y * delta_y) + (delta_z * delta_z);
           if (sphere_distance <= (m_Radius*m_Radius)){
             // check -> is the point inside the image?
@@ -203,9 +203,9 @@ void mitk::SeedsImage::AddSeedPoint(SeedsImageType* itkImage)
               iterator.SetIndex(setIndex);
               if (m_DrawState == -1 || m_DrawState == 1)
               {  
-                getIndex[2]=z + m_Radius -contIndex[2] ;
-                getIndex[1]=y + m_Radius -contIndex[1];
-                getIndex[0]=x + m_Radius -contIndex[0];
+                getIndex[2]=z + m_Radius -(int)contIndex[2] ;
+                getIndex[1]=y + m_Radius -(int)contIndex[1];
+                getIndex[0]=x + m_Radius -(int)contIndex[0];
                 float val = iterator.Get() + mask->GetPixel(getIndex) ;
                 if (val > 128) val = 128; 
                 if (val < -128) val = 128;
